// frontend/teampilot/src/app/services/api/v1/model/leaderBoardPlayerDTO.ts
/**
 * TeamPilot.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LeaderBoardPlayerDTO { 
    playerId?: string | null;
    playerName?: string | null;
    playerKills?: number;
    playerDeaths?: number;
    readonly playerKDRatio?: number;
}